Removed the following SuppressWarnings and fixed the file specific formattings based on codenarc tests.
Fixed the issues reported by following gradle command:
./gradlew codenarcMain codenarcTest

Split the PartyTests.groovy files into multiple files and moved the logical test cases in those files Communication, Misc and Address specific tests.
