+++ applications/party/testdef/PartyTests.xml
@@ -25,7 +25,16 @@
         <entity-xml action="load" entity-xml-url="component://party/testdef/data/PartyTestsData.xml"/>
     </test-case>
 
-    <test-group case-name="party-tests">
+    <test-case case-name="party-tests">
         <junit-test-suite class-name="org.apache.ofbiz.party.party.test.PartyTests"/>
-    </test-group>
+    </test-case>
+    <test-case case-name="party-address-tests">
+        <junit-test-suite class-name="org.apache.ofbiz.party.party.test.PartyAddressTests"/>
+    </test-case>
+    <test-case case-name="party-communication-event-tests">
+        <junit-test-suite class-name="org.apache.ofbiz.party.party.test.PartyCommunicationEventTests"/>
+    </test-case>
+    <test-case case-name="party-misc-tests">
+        <junit-test-suite class-name="org.apache.ofbiz.party.party.test.PartyMiscTests"/>
+    </test-case>
 </test-suite>
